+++ feature/settings/src/commonMain/kotlin/com/zrcoding/hackertab/settings/presentation/topics/SettingTopicsScreenViewModel.kt
@@ -39,6 +39,7 @@ class SettingTopicsScreenViewModel(
     }
 
     fun onChipClicked(topic: ChipData) {
+        if (_viewState.value.count { it.selected } <= 1 && topic.selected) return
         viewModelScope.launch {
             if (topic.selected) {
                 settingRepository.removeTopic(id = topic.id)
